Maken van de project pagina
Hierbij heb ik nieuwe fonts geimporteert die variabel zijn, dus makkelijk te stijlen. Daarnaast heb ik meer componenten gemaakt om het projecten component makkelijk te stijlen.

Mijn volgende stap is om de componenten eventueel herbruikbaar te maken voor andere onderdelen.
