+++ src/components/Github_Hyperlinks.svelte
@@ -1,0 +1,52 @@
+<script>
+  import { Icons } from '../index';
+  export let Github
+</script>
+
+<ul>
+    <li>
+      <a href={Github.live_link}>
+        <Icons Name='link' Width=20 Height=20 Color='var(--lightmode-color-dark)'/> 
+        Live Link
+      </a>
+    </li>
+
+    <li>
+      <a href={Github.link}>
+      <Icons Name='Github' Width=20 Height=20 Color='var(--lightmode-color-dark)'/> 
+      GitHub Link
+    </a>
+  </li>
+
+  <li>
+    <a href={Github.Project_board}>
+      <Icons Name='Github' Width=20 Height=20 Color='var(--lightmode-color-dark)'/> 
+      Project Board
+    </a>
+  </li>
+
+  <li>
+    <a href={Github.issues}>
+      <Icons Name='Github' Width=20 Height=20 Color='var(--lightmode-color-dark)'/> 
+      Issues
+    </a>
+  </li>
+
+  <li>
+    <a href={Github.readme}>
+      <Icons Name='Github' Width=20 Height=20 Color='var(--lightmode-color-dark)'/> 
+      ReadMe
+    </a>
+  </li>
+
+  <li>
+    <a href={Github.code}>
+      <Icons Name='Github' Width=20 Height=20 Color='var(--lightmode-color-dark)'/> 
+      Code
+    </a>
+  </li>
+</ul>
+
+<style>
+  
+</style>
