+++ src/index.js
@@ -1,6 +1,9 @@
 export { default as Footer } from './Layouts/Footer.svelte'
 export { default as Header } from './Layouts/Header.svelte'
-export { default as SubjectData} from './layouts/SubjectData.svelte'
+export { default as SubjectData} from './layouts/Subject/SubjectData.svelte'
+export { default as Github_Hyperlinks} from './components/Github_Hyperlinks.svelte'
+export { default as Keywords} from './components/Keywords.svelte'
+
 
 
 export { default as Heading } from './components/Heading.svelte'
